Further refactored device class to hold all devices specific implementations
