+++ clearblade/cloud/iot_v1/__init__.py
@@ -1,5 +1,5 @@
 from clearblade.cloud.iot_v1.client import DeviceManagerClient
-from clearblade.cloud.iot_v1.request_response import SendCommandToDeviceRequest
+from clearblade.cloud.iot_v1.devices import SendCommandToDeviceRequest
 
 __all__ = ("DeviceManagerClient",
            "SendCommandToDeviceRequest")
